Using torchao for quantization
Summary:
This PR added int4, int4-gptq, int8, 8da4w quantization to llama-fast

Test Plan:
Note: int4wo, int4wo-gptq only works for llama2 7b right now since stories has dimension 288 which needs some padding for it to work with int4 kernels

python generate.py --compile --checkpoint-path ${MODEL_PATH} --prompt "Hello, my name is" --device cuda --dtype bf16
python generate.py --compile --checkpoint-path ${MODEL_PATH} --prompt "Hello, my name is" --device cuda --quantize '{"linear:int8": {}}'
python generate.py --compile --checkpoint-path ${MODEL_PATH} --prompt "Hello, my name is" --device cuda --quantize '{"linear:int4": {"groupsize": 32}}' --dtype bf16
python generate.py --compile --checkpoint-path ${MODEL_PATH} --prompt "Hello, my name is" --device cuda --quantize '{"linear:int4-gptq": {"groupsize": 32}}' --dtype bf16
python generate.py --compile --checkpoint-path ${MODEL_PATH} --prompt "Hello, my name is" --device cuda --quantize '{"linear:8da4w": {"groupsize": 32}}' --dtype fp32

Reviewers:

Subscribers:

Tasks:

Tags:

+++ model_aoti.py
@@ -44,7 +44,7 @@ class DSOModel(nn.Module):
 
         # build transformer model
         global src, dso_src
-        
+
         src = src.replace('***my_model.so***', str(dso_path))
         async_compile = AsyncCompile()
         self.transformer_model = async_compile.cpp_pybinding(
@@ -63,5 +63,5 @@ class DSOModel(nn.Module):
         self.transformer_model(x, input_pos, logits)
         return logits
 
-    def setup_caches(self, max_batch_size, max_seq_length):
+    def setup_caches(self, max_batch_size, max_seq_length, dtype=torch.float):
         pass
